+++ pkgs/intl4x/lib/src/collator/collator.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../intl4x_test_checker.dart';
+import '../test_checker.dart';
 import '../options.dart';
 import 'collator_options.dart';
